app/layers/pieriandx_tools_layer/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pieriandx_tools"
version = "0.0.1"
description = "PierianDx Pipeline Lambda Layers"
license = "GPL-3.0-or-later"
authors = [
    "Alexis Lucattini"
]
homepage = "https://github.com/orcabus/service-pieriandx-pipeline-manager"
repository = "https://github.com/orcabus/service-pieriandx-pipeline-manager"

[tool.poetry.dependencies]
python = "^3.12, <3.15"
boto3 = "^1.28"
botocore = "^1.31"
aws_requests_auth = "^0.4.3"
v2_samplesheet_maker = "^4.2.4"
# Git dependencies
pyriandx = { git = "https://github.com/umccr/pyriandx.git", branch = "0.4.0" }
pandas = "^2.2.2"
urllib3 = "^1.26"
setuptools = "^80.0.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pyarrow = "^15.0.0"  # Pandas throws a warning if this is not installed
pytest = "^7.0.0"  # For testing only
# For typehinting only, not required at runtime
mypy-boto3-ssm = "^1.34"
mypy-boto3-s3 = "^1.34"
mypy-boto3-secretsmanager = "^1.34"
mypy-boto3-stepfunctions = "^1.34"
mypy-boto3-lambda = "^1.34"
